Fixed Addressing
Made an oopsie on the addressing for the modbus connection.  was using the boolean address 0-2000 which is incorrect. changed them to the modbus input address 0-1000.

i updated the HMI addresses in the Excel doc

TODO: add buttons for going between monitor pages and going to the control window for that specific module.