Fall 2022 HMI for demoing factory in Moscow
Edited the existing HMI to enable a demo mode. In the demo mode users can select a puck from a GUI. Once selected(and in auto mode) the factory will autonomously pick up and process the puck from start to finish. Flipping auto mode in manual acts as an Emergency stop as well as enabling full manual control of the factory. Flipping back to auto should reset everything back to a known good state.(reset positions, flip all switches that are on to off, etc.) This is meant to be used in tandem with the new Moscow Fa22 Demo